Major update to project structure!
After watching a talk on using traits in rust, I decided to rewrite endpoints to use a common EndpointResult trait. This simplifies the API struct itself,
and moves more of the burden of doing stuff onto the different endpoint files, which is great!

In addition, this library now has the ability to parse "error" json results from 200 OK calls, like the `{ error: 'message' }` format that the server returns for invalid room, etc.
Invalid room messages in particular will result in a unique error enum result!

With this, I've also moved some of the unauthenticated tests to unauthenticated_tests.rs, and added a test there for room_terrain() on an invalid room returning the specific ApiError::InvalidRoom result!