Add rudementary room detail parsing.
Right now this kind of cheats and deserializes RoomObjects as a serde_json::Value, but I believe this is the sanest course of action.

It would be possible to have an enum of different kind of room objects if we knew what 'type' they are, but the type is only included in the first update where each object is introduced. From then on, only ids are given to identify them (and only changes are included).

Next I think it would be important to add a 'DescribedObject' enum which provides a `new(serde_json::Value) -> Self` method, as well as an `update_with(serde_json::Value)` method for updates after that. It would be up to the user of screeps-api to store these objects somewhere, and to link up the ids to provide them with updates.