Update to hyper 1.12 and switch to permanent auth tokens.
This does two overhauls at once, which is bad, but eh. It's a major
verison upgrade for our biggest dependency, and an upgade of the API -
screeps now uses captchas on login, and has introduced permanent API
tokens for alternative clients to use.

The removal of temp auth token management has been mostly for the sake
of simplicity. I mean we'll have to add it back in, but eh, that's
later. I'm pretty sure no one's seriously using this library anyways,
so.