Replace broken custom Deserialize impl with derive
I'm really not sure what compelled me to originally write this in
6fd385abf83a9de00e51d939e7a07283d1cfe4aa. I'm 100% sure this is a copy
of the derived code with `#[serde(untagged)]`, and I think when I do
this sort of thing I'd usually edit something after copying it. But I
can't find anything changed, and the derive seems to work fine?

In any case, the handwritten code broke with the latest serde update
because it referred to private inner parts, and that's not great.